requirements.txt
google-generativeai==0.3.2
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
webdriver-manager==4.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3